requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-groq>=0.1.5
langchain-community>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
groq>=0.8.0
